+++ src/main/java/duke/task/Deadline.java
@@ -1,7 +1,10 @@
+package duke.task;
+
+import duke.parser.Parser;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.text.SimpleDateFormat;
 import java.util.Locale;
 
 public class Deadline extends Task {
